pyproject.toml
[project]
name = "new-mm-py"
version = "0.1.0"
description = "Martingale Grid Bot for multiple exchanges"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aster-connector-python",
    "avantis-trader-sdk>=0.2.1",
    "hyperliquid-python-sdk>=0.19.0",
    "lighter-sdk",
    "pybit>=5.6.0",
    "paradex-py==0.5.0rc2",
    "python-dotenv>=1.1.1",
    "grvt-pysdk>=0.1.19",
]

[tool.uv.sources]
aster-connector-python = { git = "https://github.com/asterdex/aster-connector-python.git" }
lighter-sdk = { git = "https://github.com/elliottech/lighter-python.git" }

[tool.uv]
override-dependencies = [
  "eth-account>=0.13.4",
  "hexbytes>=1.2.0"
]
